pyproject.toml
[tool.poetry]
name = "lumivance"
version = "0.1.0"
description = "The Digital Nervous System for the novel Lumivance."
authors = ["Ashton Coghlan <ashton@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0.0"
chromadb = "^0.4.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
